feat(js): move workspace tests
