+++ e2e/js/src/js.test.ts
@@ -1,10 +1,12 @@
 import {
+  checkFilesDoNotExist,
+  checkFilesExist,
   cleanupProject,
   createFile,
-  expectJestTestsToPass,
   newProject,
   readJson,
   runCLI,
+  runCLIAsync,
   uniq,
   updateFile,
   updateJson,
@@ -87,7 +89,80 @@ export function ${lib}Wildcard() {
     );
   }, 240_000);
 
-  it('should run default jest tests', async () => {
-    await expectJestTestsToPass('@nrwl/js:lib');
-  }, 240_000);
+  it('should create a library that can be linted and tested', async () => {
+    const libName = uniq('mylib');
+    const dirName = uniq('dir');
+
+    runCLI(`generate @nrwl/js:lib ${libName} --directory ${dirName}`);
+
+    checkFilesExist(
+      `libs/${dirName}/${libName}/src/index.ts`,
+      `libs/${dirName}/${libName}/README.md`
+    );
+
+    // Lint
+    const result = runCLI(`lint ${dirName}-${libName}`);
+
+    expect(result).toContain(`Linting "${dirName}-${libName}"...`);
+    expect(result).toContain('All files pass linting.');
+
+    // Test
+    const testResult = await runCLIAsync(`test ${dirName}-${libName}`);
+    expect(testResult.combinedOutput).toContain(
+      'Test Suites: 1 passed, 1 total'
+    );
+  }, 500_000);
+
+  it('should be able to use and be used by other libs', () => {
+    const consumerLib = uniq('consumer');
+    const producerLib = uniq('producer');
+
+    runCLI(`generate @nrwl/js:lib ${consumerLib} --bundler=none`);
+    runCLI(`generate @nrwl/js:lib ${producerLib} --bundler=none`);
+
+    updateFile(
+      `libs/${producerLib}/src/lib/${producerLib}.ts`,
+      'export const a = 0;'
+    );
+
+    updateFile(
+      `libs/${consumerLib}/src/lib/${consumerLib}.ts`,
+      `
+    import { a } from '@${scope}/${producerLib}';
+
+    export function ${consumerLib}() {
+      return a + 1;
+    }`
+    );
+    updateFile(
+      `libs/${consumerLib}/src/lib/${consumerLib}.spec.ts`,
+      `
+    import { ${consumerLib} } from './${consumerLib}';
+
+    describe('', () => {
+      it('should return 1', () => {
+        expect(${consumerLib}()).toEqual(1);
+      });
+    });`
+    );
+
+    runCLI(`test ${consumerLib}`);
+  });
+
+  it('should not be able to be built when it has no bundler', () => {
+    const nonBuildable = uniq('buildable');
+
+    runCLI(`generate @nrwl/js:lib ${nonBuildable} --bundler=none`);
+
+    const result = runCLI(`build ${nonBuildable}`);
+
+    expect(result).toContain(
+      `Compiling TypeScript files for project "${nonBuildable}"...`
+    );
+    expect(result).toContain(
+      `Done compiling TypeScript files for project "${nonBuildable}".`
+    );
+
+    checkFilesDoNotExist(`dist/libs/${nonBuildable}/README.md`);
+  });
 });
